pyproject.toml
[project]
name = "note-summarizer-api"
version = "0.1.0"
description = ""
authors = [
    {name = "max0021",email = "541451@student.saxion.nl"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "newspaper3k (>=0.2.8,<0.3.0)",
    "lxml-html-clean (>=0.4.2,<0.5.0)",
    "openai (>=1.88.0,<2.0.0)",
    "transformers (>=4.52.4,<5.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "nltk (>=3.9.1,<4.0.0)",
    "trafilatura (>=2.0.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

